move quote magic on builder to static classe
